refactor out ModuleNode as a react component
