+++ apps/web/flow/dag.tsx
@@ -1,23 +1,4 @@
-import { Node, Edge, Position, Handle } from 'react-flow-renderer'
-
-export function ModuleNode({ data }: any) {
-  return (
-    <div>
-      <div className="bg-white flex flex-col justify-center h-24 w-40 shadow-md rounded-md border-2 border-gray-300">
-        <Handle type="target" position={Position.Left} />
-        <Handle type="source" position={Position.Right} />
-        <div className="h-min">
-          <div className="text-2xl font-semibold text-gray-700 text-center">
-            {data.moduleCode}
-          </div>
-          <div className="text-sm font-medium tracking-tight text-gray-400 text-center leading-5">
-            {data.title}
-          </div>
-        </div>
-      </div>
-    </div>
-  )
-}
+import { Node, Edge } from 'react-flow-renderer'
 
 const modules = [
   {
