+++ apps/web/pages/index.tsx
@@ -1,73 +1,57 @@
-import { ReactNode, useEffect, useState } from 'react'
-import Search from '@/components/Search'
-import { ModuleCondensed } from 'database'
-import { ResultDisplay } from '@/components/module-search/Results'
-import { SelectedDisplay } from '@/components/module-search/Selected'
+import { useCallback, useEffect, useState } from 'react'
+import ReactFlow, {
+  Controls,
+  applyNodeChanges,
+  applyEdgeChanges,
+} from 'react-flow-renderer'
+import { initialNodes, initialEdges } from '@/flow/dag'
+import { ModuleNode } from '@/components/flow/ModuleNode'
 import { useSelector, useDispatch } from 'react-redux'
-import { BuilderState, clearBuilderModules } from '@/store/builder'
+import { setFlowSelection, FlowState } from '@/store/flow'
 
-export default function SearchPage() {
-  const [results, setResults] = useState<ModuleCondensed[]>([])
-  const builderSelection = useSelector<BuilderState, ModuleCondensed[]>(
-    (state) => state.builder.moduleCondensed
-  )
+const nodeTypes = { moduleNode: ModuleNode }
+
+export default function Modtree() {
   const dispatch = useDispatch()
+  const treeSelection = useSelector<FlowState, string>(
+    (state) => state.flow.moduleCode
+  )
+  const [nodes, setNodes] = useState(initialNodes)
+  const [edges, setEdges] = useState(initialEdges)
+
+  const onNodesChange = useCallback(
+    (changes) => setNodes((nds) => applyNodeChanges(changes, nds)),
+    [setNodes]
+  )
+  const onEdgesChange = useCallback(
+    (changes) => setEdges((eds) => applyEdgeChanges(changes, eds)),
+    [setEdges]
+  )
 
   useEffect(() => {
-    console.log('builder selection:', builderSelection)
-  }, [builderSelection])
-
-  const ResultContainer = (props: { children: ReactNode }) => {
-    return (
-      <div className="bg-white rounded-md shadow-md w-full rounded-md overflow-hidden">
-        {props.children}
-      </div>
-    )
-  }
-
-  const SelectedContainer = (props: { children: ReactNode }) => {
-    return (
-      <div className="bg-white rounded-md shadow-md w-full overflow-y-scroll h-[40rem]">
-        {props.children}
-      </div>
-    )
-  }
+    console.log(treeSelection)
+  }, [treeSelection])
 
   return (
-    <div>
-      <h1 className="text-4xl mt-12 mb-12 font-semibold tracking-normal text-gray-700">
-        modtree
-      </h1>
-      <div className="flex flex-row justify-center overflow-y-hidden">
-        <div className="w-1/4 bg-white mr-6 rounded-md shadow-md mb-4">
-          <div className="flex flex-row">
-            <h2 className="px-4 py-3 text-xl tracking-tight font-semibold text-gray-500 flex-1">
-              Module List
-            </h2>
-            <div className="flex flex-col justify-center mr-4 tracking-normal">
-              <div
-                className="text-gray-400 rounded-md px-1.5 hover:bg-gray-200 cursor-pointer active:bg-gray-300"
-                onClick={() => dispatch(clearBuilderModules())}
-              >
-                clear
-              </div>
-            </div>
-          </div>
-          <SelectedContainer>
-            <SelectedDisplay />
-          </SelectedContainer>
-        </div>
-        <div className="mb-4 w-full max-w-xl">
-          <Search setResults={setResults} />
-          {results.length > 0 ? (
-            <div className="flex flex-row justify-center mt-6">
-              <ResultContainer>
-                <ResultDisplay results={results} />
-              </ResultContainer>
-            </div>
-          ) : null}
-        </div>
-      </div>
+    <div className="h-screen w-screen bg-gray-50">
+      <ReactFlow
+        nodes={nodes}
+        edges={edges}
+        zoomOnDoubleClick={false}
+        onNodesChange={onNodesChange}
+        onEdgesChange={onEdgesChange}
+        nodeTypes={nodeTypes}
+        fitView={true}
+        onSelectionChange={(e) => {
+          const moduleCodes = e.nodes.map((x) => x.data.moduleCode)
+          dispatch(setFlowSelection(moduleCodes))
+        }}
+        fitViewOptions={{ maxZoom: 1 }}
+        defaultZoom={1}
+        maxZoom={2}
+      >
+        <Controls />
+      </ReactFlow>
     </div>
   )
 }
