+++ apps/web/tsconfig.json
@@ -7,7 +7,8 @@
     "paths": {
       "@/components/*": ["components/*"],
       "@/store/*": ["store/*"],
-      "@/utils/*": ["utils/*"]
+      "@/utils/*": ["utils/*"],
+      "@/flow/*": ["flow/*"]
     }
   }
 }
